moved eReporting view and controller to appropriate modules and added menu item
