+++ spring/views/src/main/webapp/WEB-INF/tld/functions.tld
@@ -23,6 +23,11 @@
         <function-class>org.n52.sos.web.JstlFunctions</function-class>
         <function-signature>boolean hasAdministrator()</function-signature>
     </function>
+     <function>
+        <name>hasClass</name>
+        <function-class>org.n52.sos.web.JstlFunctions</function-class>
+        <function-signature>boolean hasClass(java.lang.String)</function-signature>
+    </function>
     <function>
         <name>viewExists</name>
         <function-class>org.n52.sos.web.JstlFunctions</function-class>
@@ -37,5 +42,5 @@
         <name>mapToJson</name>
         <function-class>org.n52.sos.web.JstlFunctions</function-class>
         <function-signature>java.lang.String mapToJson(java.util.Map)</function-signature>
-    </function>    
+    </function>
 </taglib>
