+++ spring/views/src/main/webapp/WEB-INF/views/common/header.jsp
@@ -146,6 +146,13 @@
                                                                     <span class="menu-title">I18N Settings</span>
                                                                 </a>
                                                             </li>
+                                                            <c:if test="${sos:hasClass('org.n52.sos.web.admin.AdminEReportingHeaderController')}">
+                                                                <li>
+                                                                    <a href="<c:url value="/admin/ereporting" />">
+                                                                        <span class="menu-title">eReporting</span>
+                                                                    </a>
+                                                                </li>
+                                                            </c:if>
                                                         </ul>
                                                     </li>
                                                     <li>
